+++ Rebus/Persistence/SqlServer/DbConnectionProvider.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Configuration;
 using System.Data.SqlClient;
-using System.Linq;
 using System.Threading.Tasks;
 using System.Transactions;
 using Rebus.Logging;
@@ -31,45 +30,18 @@ namespace Rebus.Persistence.SqlServer
 
             _log = rebusLoggerFactory.GetCurrentClassLogger();
 
-            var connectionString = GetConnectionString(connectionStringOrConnectionStringName);
-
-            _connectionString = EnsureMarsIsEnabled(connectionString);
+            _connectionString = GetConnectionString(connectionStringOrConnectionStringName);
 
             IsolationLevel = IsolationLevel.ReadCommitted;
-        }
-
-        string EnsureMarsIsEnabled(string connectionString)
-        {
-            var connectionStringParameters = connectionString.Split(";".ToCharArray(), StringSplitOptions.RemoveEmptyEntries)
-                .Select(kvpString =>
-                {
-                    var tokens = kvpString.Split("=".ToCharArray(), StringSplitOptions.RemoveEmptyEntries);
-
-                    return new
-                    {
-                        Key = tokens[0],
-                        Value = string.Join("=", tokens.Skip(1))
-                    };
-                })
-                .ToDictionary(a => a.Key, a => a.Value, StringComparer.InvariantCultureIgnoreCase);
-
-            if (!connectionStringParameters.ContainsKey("MultipleActiveResultSets"))
-            {
-                _log.Info("Supplied connection string does not have MARS enabled - the connection string will be modified to enable MARS!");
-                return connectionString + ";MultipleActiveResultSets=true";
-            }
-
-            return connectionString;
         }
 
         static string GetConnectionString(string connectionStringOrConnectionStringName)
         {
             var connectionStringSettings = ConfigurationManager.ConnectionStrings[connectionStringOrConnectionStringName];
 
-            if (connectionStringSettings != null)
-                return connectionStringSettings.ConnectionString;
-
-            return connectionStringOrConnectionStringName;
+            return connectionStringSettings != null 
+                ? connectionStringSettings.ConnectionString 
+                : connectionStringOrConnectionStringName;
         }
 
         /// <summary>